Add functionality for truncating posts & add hover styles
Add functionality for truncating a post on the feed when it is 375 characters or longer.
Users can view the full post by clicking on the post.
A 'read more' message is also shown for truncated posts
Add hover styles to Post component so it is more clear the user can click on it